Put the debugging command to display the key log in the Angband 4 command system.  Reversed the printed order of the entries to match what the printed label says.
