+++ src/cmd-core.h
@@ -117,6 +117,7 @@ typedef enum cmd_code {
 	CMD_WIZ_CURE_ALL,
 	CMD_WIZ_DETECT_ALL_LOCAL,
 	CMD_WIZ_DETECT_ALL_MONSTERS,
+	CMD_WIZ_DISPLAY_KEYLOG,
 	CMD_WIZ_HIT_ALL_LOS,
 	CMD_WIZ_INCREASE_EXP,
 	CMD_WIZ_JUMP_LEVEL,
