+++ src/cmd-wizard.c
@@ -37,7 +37,6 @@
 #include "ui-map.h"
 #include "ui-output.h"
 #include "ui-target.h"
-#include "ui-term.h"
 #include "wizard.h"
 
 
@@ -716,6 +715,55 @@ void do_cmd_wiz_detect_all_monsters(struct command *cmd)
 		0, 0, 0, 500, 500, NULL);
 	effect_simple(EF_DETECT_INVISIBLE_MONSTERS, source_player(), "0",
 		0, 0, 0, 500, 500, NULL);
+}
+
+
+/**
+ * Display the keycodes the user has been generating (CMD_WIZ_DISPLAY_KEYLOG).
+ * Takes no arguments from cmd.
+ */
+void do_cmd_wiz_display_keylog(struct command *cmd)
+{
+	int i;
+	char buf[50];
+	char buf2[12];
+	struct keypress keys[2] = {KEYPRESS_NULL, KEYPRESS_NULL};
+
+	screen_save();
+
+	prt("Previous keypresses (top most recent):", 0, 0);
+
+	for (i = 0; i < KEYLOG_SIZE; i++) {
+		if (i < log_size) {
+			/*
+			 * Find the keypress from the log; log_i is one past
+			 * the most recent.
+			 */
+			int j = (log_i > i) ?
+				log_i - i - 1 : log_i - i - 1 + KEYLOG_SIZE;
+			struct keypress k = keylog[j];
+
+			/*
+			 * ugh. it would be nice if there was a version of
+			 * keypress_to_text which took only one keypress.
+			 */
+			keys[0] = k;
+			keypress_to_text(buf2, sizeof(buf2), keys, true);
+
+			/* format this line of output */
+			strnfmt(buf, sizeof(buf), "    %-12s (code=%u mods=%u)",
+				buf2, k.code, k.mods);
+		} else {
+			/* create a blank line of output */
+			strnfmt(buf, sizeof(buf), "%40s", "");
+		}
+
+		prt(buf, i + 1, 0);
+	}
+
+	prt("Press any key to continue.", KEYLOG_SIZE + 1, 0);
+	anykey();
+	screen_load();
 }
